pyproject.toml
[project]
name = "audio-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "faster-whisper>=1.0.3",
    "pyannote-audio>=3.3.1",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "librosa>=0.11.0",
    "soundfile>=0.13.1",
]
